accounts/urls.py
from django.conf.urls import include, url
from rest_framework.routers import DefaultRouter

from accounts import views

# Create a router and register our viewsets with it.
router = DefaultRouter()
router.register(r'users', views.UserViewSet)

# The API URLs are now determined automatically by the router.
urlpatterns = [
    url(r'^', include(router.urls))
]